Changed append_array to make use of already existing resize/dims functions instead.
Modified resize and tests to treat Eigen types the same as dims does (Issue 481)
