+++ test/unit/math/fwd/mat/fun/resize_test.cpp
@@ -7,12 +7,12 @@ using stan::math::fvar;
 
 TEST(AgradFwdMatrixResize, fvar_double) {
   fvar<double> x = 5;
-  std::vector<size_t> dims;
+  std::vector<int> dims;
   stan::math::resize(x,dims);
 }
 TEST(AgradFwdMatrixResize, svec_fvar_double) {
   std::vector<fvar<double> > y;
-  std::vector<size_t> dims;
+  std::vector<int> dims;
   EXPECT_EQ(0U, y.size());
 
   dims.push_back(4U);
@@ -25,10 +25,11 @@ TEST(AgradFwdMatrixResize, svec_fvar_double) {
 }
 TEST(AgradFwdMatrixResize, vec_fvar_double) {
   Matrix<fvar<double>,Dynamic,1> v(2);
-  std::vector<size_t> dims;
+  std::vector<int> dims;
   EXPECT_EQ(2, v.size());
 
   dims.push_back(17U);
+  dims.push_back(1U);
   stan::math::resize(v,dims);
   EXPECT_EQ(17, v.size());
 
@@ -38,20 +39,21 @@ TEST(AgradFwdMatrixResize, vec_fvar_double) {
 }
 TEST(AgradFwdMatrixResize, rvec_fvar_double) {
   Matrix<fvar<double>,1,Dynamic> rv(2);
-  std::vector<size_t> dims;
+  std::vector<int> dims;
   EXPECT_EQ(2, rv.size());
 
+  dims.push_back(1U);
   dims.push_back(17U);
   stan::math::resize(rv,dims);
   EXPECT_EQ(17, rv.size());
 
-  dims[0] = 3U;
+  dims[1] = 3U;
   stan::math::resize(rv,dims);
   EXPECT_EQ(3, rv.size());
 }
 TEST(AgradFwdMatrixResize, mat_fvar_double) {
   Matrix<fvar<double>,Dynamic,Dynamic> m(2,3);
-  std::vector<size_t> dims;
+  std::vector<int> dims;
   EXPECT_EQ(2, m.rows());
   EXPECT_EQ(3, m.cols());
 
@@ -64,7 +66,7 @@ TEST(AgradFwdMatrixResize, mat_fvar_double) {
 TEST(AgradFwdMatrixResize, svec_svec_fvar_double) {
   std::vector<std::vector<fvar<double> > > xx;
   EXPECT_EQ(0U,xx.size());
-  std::vector<size_t> dims;
+  std::vector<int> dims;
   dims.push_back(4U);
   dims.push_back(5U);
   stan::math::resize(xx,dims);
@@ -80,9 +82,10 @@ TEST(AgradFwdMatrixResize, svec_svec_fvar_double) {
 TEST(AgradFwdMatrixResize, svec_v_fvar_double) {
   std::vector<Matrix<fvar<double>,Dynamic,1> > xx;
   EXPECT_EQ(0U,xx.size());
-  std::vector<size_t> dims;
+  std::vector<int> dims;
   dims.push_back(4U);
   dims.push_back(5U);
+  dims.push_back(1U);
   stan::math::resize(xx,dims);
   EXPECT_EQ(4U,xx.size());
   EXPECT_EQ(5,xx[0].size());
@@ -96,22 +99,23 @@ TEST(AgradFwdMatrixResize, svec_v_fvar_double) {
 TEST(AgradFwdMatrixResize, svec_rv_fvar_double) {
   std::vector<Matrix<fvar<double>,1,Dynamic> > xx;
   EXPECT_EQ(0U,xx.size());
-  std::vector<size_t> dims;
+  std::vector<int> dims;
   dims.push_back(4U);
+  dims.push_back(1U);
   dims.push_back(5U);
   stan::math::resize(xx,dims);
   EXPECT_EQ(4U,xx.size());
   EXPECT_EQ(5,xx[0].size());
 
   dims[0] = 3U;
-  dims[1] = 7U;
+  dims[2] = 7U;
   stan::math::resize(xx,dims);
   EXPECT_EQ(3U,xx.size());
   EXPECT_EQ(7,xx[1].size());  
 }
 TEST(AgradFwdMatrixResize, svec_svec_matrix_fvar_double) {
   std::vector<std::vector<Matrix<fvar<double>,Dynamic,Dynamic> > > mm;
-  std::vector<size_t> dims;
+  std::vector<int> dims;
   dims.push_back(4U);
   dims.push_back(5U);
   dims.push_back(6U);
@@ -124,12 +128,12 @@ TEST(AgradFwdMatrixResize, svec_svec_matrix_fvar_double) {
 }
 TEST(AgradFwdMatrixResize, fvar_fvar_double) {
   fvar<fvar<double> > x = 5;
-  std::vector<size_t> dims;
+  std::vector<int> dims;
   stan::math::resize(x,dims);
 }
 TEST(AgradFwdMatrixResize, svec_fvar_fvar_double) {
   std::vector<fvar<fvar<double> > > y;
-  std::vector<size_t> dims;
+  std::vector<int> dims;
   EXPECT_EQ(0U, y.size());
 
   dims.push_back(4U);
@@ -142,10 +146,11 @@ TEST(AgradFwdMatrixResize, svec_fvar_fvar_double) {
 }
 TEST(AgradFwdMatrixResize, vec_fvar_fvar_double) {
   Matrix<fvar<fvar<double> >,Dynamic,1> v(2);
-  std::vector<size_t> dims;
+  std::vector<int> dims;
   EXPECT_EQ(2, v.size());
 
   dims.push_back(17U);
+  dims.push_back(1U);
   stan::math::resize(v,dims);
   EXPECT_EQ(17, v.size());
 
@@ -155,20 +160,21 @@ TEST(AgradFwdMatrixResize, vec_fvar_fvar_double) {
 }
 TEST(AgradFwdMatrixResize, rvec_fvar_fvar_double) {
   Matrix<fvar<fvar<double> >,1,Dynamic> rv(2);
-  std::vector<size_t> dims;
+  std::vector<int> dims;
   EXPECT_EQ(2, rv.size());
 
+  dims.push_back(1U);
   dims.push_back(17U);
   stan::math::resize(rv,dims);
   EXPECT_EQ(17, rv.size());
 
-  dims[0] = 3U;
+  dims[1] = 3U;
   stan::math::resize(rv,dims);
   EXPECT_EQ(3, rv.size());
 }
 TEST(AgradFwdMatrixResize, mat_fvar_fvar_double) {
   Matrix<fvar<fvar<double> >,Dynamic,Dynamic> m(2,3);
-  std::vector<size_t> dims;
+  std::vector<int> dims;
   EXPECT_EQ(2, m.rows());
   EXPECT_EQ(3, m.cols());
 
@@ -181,7 +187,7 @@ TEST(AgradFwdMatrixResize, mat_fvar_fvar_double) {
 TEST(AgradFwdMatrixResize, svec_svec_fvar_fvar_double) {
   std::vector<std::vector<fvar<fvar<double> > > > xx;
   EXPECT_EQ(0U,xx.size());
-  std::vector<size_t> dims;
+  std::vector<int> dims;
   dims.push_back(4U);
   dims.push_back(5U);
   stan::math::resize(xx,dims);
@@ -197,9 +203,10 @@ TEST(AgradFwdMatrixResize, svec_svec_fvar_fvar_double) {
 TEST(AgradFwdMatrixResize, svec_v_fvar_fvar_double) {
   std::vector<Matrix<fvar<fvar<double> >,Dynamic,1> > xx;
   EXPECT_EQ(0U,xx.size());
-  std::vector<size_t> dims;
+  std::vector<int> dims;
   dims.push_back(4U);
   dims.push_back(5U);
+  dims.push_back(1U);
   stan::math::resize(xx,dims);
   EXPECT_EQ(4U,xx.size());
   EXPECT_EQ(5,xx[0].size());
@@ -213,22 +220,23 @@ TEST(AgradFwdMatrixResize, svec_v_fvar_fvar_double) {
 TEST(AgradFwdMatrixResize, svec_rv_fvar_fvar_double) {
   std::vector<Matrix<fvar<fvar<double> >,1,Dynamic> > xx;
   EXPECT_EQ(0U,xx.size());
-  std::vector<size_t> dims;
+  std::vector<int> dims;
   dims.push_back(4U);
+  dims.push_back(1U);
   dims.push_back(5U);
   stan::math::resize(xx,dims);
   EXPECT_EQ(4U,xx.size());
   EXPECT_EQ(5,xx[0].size());
 
   dims[0] = 3U;
-  dims[1] = 7U;
+  dims[2] = 7U;
   stan::math::resize(xx,dims);
   EXPECT_EQ(3U,xx.size());
   EXPECT_EQ(7,xx[1].size());  
 }
 TEST(AgradFwdMatrixResize, svec_svec_matrix_fvar_fvar_double) {
   std::vector<std::vector<Matrix<fvar<fvar<double> >,Dynamic,Dynamic> > > mm;
-  std::vector<size_t> dims;
+  std::vector<int> dims;
   dims.push_back(4U);
   dims.push_back(5U);
   dims.push_back(6U);
