+++ test/unit/math/prim/mat/fun/resize_test.cpp
@@ -10,12 +10,12 @@ using stan::math::row_vector_d;
 
 TEST(MathMatrix, resize_double) {
   double x = 5;
-  std::vector<size_t> dims;
+  std::vector<int> dims;
   stan::math::resize(x,dims);
 }
 TEST(MathMatrix, resize_svec_double) {
   std::vector<double> y;
-  std::vector<size_t> dims;
+  std::vector<int> dims;
   EXPECT_EQ(0U, y.size());
 
   dims.push_back(4U);
@@ -28,10 +28,11 @@ TEST(MathMatrix, resize_svec_double) {
 }
 TEST(MathMatrix, resize_vec_double) {
   Matrix<double,Dynamic,1> v(2);
-  std::vector<size_t> dims;
+  std::vector<int> dims;
   EXPECT_EQ(2, v.size());
 
   dims.push_back(17U);
+  dims.push_back(1U);
   stan::math::resize(v,dims);
   EXPECT_EQ(17, v.size());
 
@@ -41,20 +42,21 @@ TEST(MathMatrix, resize_vec_double) {
 }
 TEST(MathMatrix, resize_rvec_double) {
   Matrix<double,1,Dynamic> rv(2);
-  std::vector<size_t> dims;
+  std::vector<int> dims;
   EXPECT_EQ(2, rv.size());
 
+  dims.push_back(1U);
   dims.push_back(17U);
   stan::math::resize(rv,dims);
   EXPECT_EQ(17, rv.size());
 
-  dims[0] = 3U;
+  dims[1] = 3U;
   stan::math::resize(rv,dims);
   EXPECT_EQ(3, rv.size());
 }
 TEST(MathMatrix, resize_mat_double) {
   Matrix<double,Dynamic,Dynamic> m(2,3);
-  std::vector<size_t> dims;
+  std::vector<int> dims;
   EXPECT_EQ(2, m.rows());
   EXPECT_EQ(3, m.cols());
 
@@ -67,7 +69,7 @@ TEST(MathMatrix, resize_mat_double) {
 TEST(MathMatrix, resize_svec_svec_double) {
   std::vector<std::vector<double> > xx;
   EXPECT_EQ(0U,xx.size());
-  std::vector<size_t> dims;
+  std::vector<int> dims;
   dims.push_back(4U);
   dims.push_back(5U);
   stan::math::resize(xx,dims);
@@ -83,9 +85,10 @@ TEST(MathMatrix, resize_svec_svec_double) {
 TEST(MathMatrix, resize_svec_v_double) {
   std::vector<Matrix<double,Dynamic,1> > xx;
   EXPECT_EQ(0U,xx.size());
-  std::vector<size_t> dims;
+  std::vector<int> dims;
   dims.push_back(4U);
   dims.push_back(5U);
+  dims.push_back(1U);
   stan::math::resize(xx,dims);
   EXPECT_EQ(4U,xx.size());
   EXPECT_EQ(5,xx[0].size());
@@ -99,22 +102,23 @@ TEST(MathMatrix, resize_svec_v_double) {
 TEST(MathMatrix, resize_svec_rv_double) {
   std::vector<Matrix<double,1,Dynamic> > xx;
   EXPECT_EQ(0U,xx.size());
-  std::vector<size_t> dims;
+  std::vector<int> dims;
   dims.push_back(4U);
+  dims.push_back(1U);
   dims.push_back(5U);
   stan::math::resize(xx,dims);
   EXPECT_EQ(4U,xx.size());
   EXPECT_EQ(5,xx[0].size());
 
   dims[0] = 3U;
-  dims[1] = 7U;
+  dims[2] = 7U;
   stan::math::resize(xx,dims);
   EXPECT_EQ(3U,xx.size());
   EXPECT_EQ(7,xx[1].size());  
 }
 TEST(MathMatrix, resize_svec_svec_matrix_double) {
   std::vector<std::vector<Matrix<double,Dynamic,Dynamic> > > mm;
-  std::vector<size_t> dims;
+  std::vector<int> dims;
   dims.push_back(4U);
   dims.push_back(5U);
   dims.push_back(6U);
